pgdatahub/requirements.txt
# PGDataHub ETL Requirements
# Python 3.10+

# Database
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0

# Excel Processing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Configuration
pyyaml>=6.0

# CLI
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
